+++ multiboot2/Changelog.md
@@ -1,6 +1,6 @@
 # CHANGELOG for crate `multiboot2`
 
-## 0.20.1
+## 0.20.1 (2024-05-26)
 
 - fixed the handling of `EFIMemoryMapTag` and `EFIMemoryAreaIter`
 - **BREAKING** Fixed wrong creation of `EFIMemoryMapTag`.